# src/main/resources/openapi/documentation.yaml
openapi: 3.0.3
info:
  title: FinGenius
  description: Financial Application (Backend)
  version: 1.0.0
servers:
- url: /
paths:
...
  /invoices:
    get:
      description: Route to get invoice(s)
      parameters:
      - name: type
        in: query
        required: false
        schema:
          type: string
      - name: id
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceJson'
        '404':
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
              examples:
                Example#1:
                  value: Invoice not found.
                Example#2:
                  value: Partner not found.
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
              examples:
                Example#1:
                  value: Failed to retrieve invoice(s).
    post:
      description: Route to add an invoice
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Invoice'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                type: string
              examples:
                Example#1:
                  value: Invoice added successfully.
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
              examples:
                Example#1:
                  value: Failed to add invoice.
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: string
              examples:
                Example#1:
                  value: Invalid data format.
...
components:
  schemas:
...
    InvoiceItem:
      type: object
      properties:
        productName:
          type: string
        quantity:
          type: string
        rate:
          type: string
        valueOfSupplies:
          type: string
        salesTax:
          type: string
        netAmount:
          type: string
    Invoice:
      type: object
      properties:
        _id:
          type: string
        type:
          type: string
        invoiceDate:
          type: string
        dueDate:
          type: string
        invoiceNo:
          type: string
        partnerId:
          type: string
        creditTerm:
          type: string
        reference:
          type: string
        invoiceTotal:
          type: string
        invoiceItems:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItem'
    Partner:
      type: object
      properties:
        _id:
          type: string
        type:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        status:
          type: string
        contactNo:
          type: string
        openingBalance:
          type: string
        address:
          type: string
        repName:
          type: string
        repContact:
          type: string
        repDesignation:
          type: string
...
    InvoiceJson:
      type: object
      properties:
        partner:
          $ref: '#/components/schemas/Partner'
        invoice:
          $ref: '#/components/schemas/Invoice'